Fixed warning

+++ src/DS1624.cpp
@@ -92,7 +92,7 @@ float DS1624::GetTemperature()
       // when conversion done
       Wire.beginTransmission(_address);
       Wire.write(ACCESS_CONFIG);
-      Wire.requestFrom(_address, 1, true);
+      Wire.requestFrom(_address, (size_t)1, true);
       
       // Wait for data sent from sensor
       while(!Wire.available());
@@ -105,7 +105,7 @@ float DS1624::GetTemperature()
   // Request to read last converted temperature value
   Wire.beginTransmission(_address);
   Wire.write(READ_TEMPERATURE);
-  Wire.requestFrom(_address, 2, true);
+  Wire.requestFrom(_address, (size_t)2, true);
 
   // Wait for data sent from sensor
   while(!Wire.available());
